[refactor] #110 공통 facade 패키지에서 comment 내부 facade 패키지로 이동
