+++ src/main/java/com/ggang/be/api/comment/facade/CommentFacade.java
@@ -1,4 +1,4 @@
-package com.ggang.be.api.facade;
+package com.ggang.be.api.comment.facade;
 
 import com.ggang.be.api.comment.dto.ReadCommentRequest;
 import com.ggang.be.api.comment.dto.ReadCommentResponse;
@@ -6,8 +6,8 @@ import com.ggang.be.api.comment.dto.WriteCommentEntityDto;
 import com.ggang.be.api.comment.dto.WriteCommentRequest;
 import com.ggang.be.api.comment.dto.WriteCommentResponse;
 import com.ggang.be.api.comment.service.CommentService;
-import com.ggang.be.api.group.CommentFacadeHandler;
-import com.ggang.be.api.group.CommentRegistry;
+import com.ggang.be.api.comment.registry.CommentFacadeHandler;
+import com.ggang.be.api.comment.registry.CommentRegistry;
 import com.ggang.be.api.user.service.UserService;
 import com.ggang.be.domain.comment.CommentEntity;
 import com.ggang.be.domain.user.UserEntity;
