+++ src/main/java/com/ggang/be/api/comment/facade/EveryGroupCommentFacadeHandler.java
@@ -1,11 +1,11 @@
-package com.ggang.be.api.facade;
+package com.ggang.be.api.comment.facade;
 
 import com.ggang.be.api.comment.dto.ReadCommentRequest;
 import com.ggang.be.api.comment.dto.ReadCommentResponse;
 import com.ggang.be.api.comment.dto.WriteCommentEntityDto;
 import com.ggang.be.api.comment.dto.WriteCommentRequest;
 import com.ggang.be.api.comment.dto.WriteCommentResponse;
-import com.ggang.be.api.group.CommentFacadeHandler;
+import com.ggang.be.api.comment.registry.CommentFacadeHandler;
 import com.ggang.be.api.group.everyGroup.service.EveryGroupService;
 import com.ggang.be.api.userEveryGroup.service.UserEveryGroupService;
 import com.ggang.be.domain.comment.CommentEntity;
